requirements.txt
streamlit==1.32.0
supabase==2.3.0
anthropic==0.20.0
python-dotenv==1.0.0
pandas==2.2.0
Pillow==10.2.0